xtask: split audit.rs into audit/ submodules
Pure code move: mod.rs (classification entry points), signatures.rs
(K1/K2/K3 detectors), interactive.rs (review/spot-audit), batch.rs
(emit/ingest/amend), report.rs (render_report and its split-out
sections), fixtures.rs. render_report's clippy::too_many_lines split
into five ordered helpers (tally_by_stratum, stratum_table_lines,
k_sensitivity_lines, flagged_lines, out_of_scope_lines); ratchet row
and #[allow] removed.
