+++ scripts/ratchet.txt
@@ -6,7 +6,6 @@
 # lint	file	function
 clippy::cognitive_complexity	mandible-extract/src/help_text/sections/scan.rs	scan_headingless_invocation_table
 clippy::cognitive_complexity	xtask/src/transition.rs	render_markdown
-clippy::too_many_lines	xtask/src/audit.rs	render_report
 clippy::too_many_lines	xtask/src/main.rs	run_coverage
 clippy::too_many_lines	xtask/src/single_dash_long.rs	self_checks
 clippy::too_many_lines	xtask/src/transition.rs	render_markdown
