+++ xtask/src/audit/fixtures.rs
@@ -1,0 +1,218 @@
+//! `xtask audit fixtures`: turn every reviewed (non-`skip`) entry into a
+//! `corpus/`-shaped fixture — `correct` gets a real `expected.snap`;
+//! `incomplete`/`wrong` get `[xfail]` with the reviewer's note as `reason`.
+
+use super::classify_one;
+use super::interactive::render_snapshot;
+use crate::status;
+use mandible_core::audit::{load, verdict_path};
+use mandible_core::CommandNode;
+use std::path::Path;
+
+/// `xtask audit fixtures`: turn every reviewed (non-`skip`) entry into a
+/// `corpus/README.md`-shaped fixture directory under `corpus_dir` — capture
+/// files, a pre-filled `meta.toml`, and (for `correct`) an `expected.snap`.
+///
+/// Stages by default (`corpus_dir` defaults to `<dir>/<seed>/fixtures`,
+/// not `corpus/`): a `wrong`/`incomplete` verdict becomes an `[xfail]`
+/// with no currently-failing `[contract]` field, which `xtask corpus`
+/// otherwise reads as "the bug appears fixed" and vacuously passes.
+/// Writes the reviewer's note as `reason` plus `expected_framework`
+/// (Tier A′'s detection, not a correctness claim). Promoting a staged
+/// fixture into `corpus/` is a deliberate act, same spirit as `--bless`.
+///
+/// A `correct` verdict gets a real `expected.snap` and may ship green
+/// immediately, wherever `corpus_dir` points.
+pub fn cmd_fixtures(
+    dir: &Path,
+    seed: u64,
+    corpus_dir: &Path,
+    only: Option<Vec<String>>,
+    force: bool,
+) -> anyhow::Result<()> {
+    let path = verdict_path(dir, seed);
+    let file = load(&path)?;
+    let mut written = 0usize;
+    let mut skipped_no_capture = 0usize;
+    let mut skipped_verdict = 0usize;
+    let mut skipped_exists = 0usize;
+
+    for entry in &file.entries {
+        // Reads the effective (post-amendment) verdict and note: a fixture
+        // generated from an entry that was amended after review must
+        // reflect the corrected truth, not the superseded original — the
+        // whole point of an amendment is that the project's belief about
+        // the tool changed, and a fixture is exactly the kind of durable
+        // artifact that would otherwise silently encode the old, wrong
+        // belief forever.
+        let Some(verdict) = entry.effective_verdict() else {
+            continue;
+        };
+        let note = entry.effective_note();
+        if verdict == "skip" {
+            skipped_verdict += 1;
+            continue;
+        }
+        if let Some(only) = &only {
+            if !only.iter().any(|t| t == &entry.tool) {
+                continue;
+            }
+        }
+
+        let fixture_dir = corpus_dir
+            .join(&entry.tool)
+            .join(format!("audit-seed{seed}"));
+        if fixture_dir.exists() && !force {
+            println!(
+                "{}: {} already exists — pass --force to overwrite (review the existing \
+                 fixture for machine-specific content first, corpus/README.md step 3)",
+                entry.tool,
+                fixture_dir.display()
+            );
+            skipped_exists += 1;
+            continue;
+        }
+
+        let classified = classify_one(&entry.tool);
+        let Some((argv_tail, output)) = classified.raw_capture else {
+            println!(
+                "{}: no raw capture available, skipping fixture emission",
+                entry.tool
+            );
+            skipped_no_capture += 1;
+            continue;
+        };
+
+        std::fs::create_dir_all(&fixture_dir)
+            .map_err(|e| anyhow::anyhow!("creating {}: {e}", fixture_dir.display()))?;
+        std::fs::write(fixture_dir.join("help.txt"), &output.stdout)?;
+        if !output.stderr.is_empty() {
+            std::fs::write(fixture_dir.join("help.stderr.txt"), &output.stderr)?;
+        }
+
+        let mut argv = vec![entry.tool.clone()];
+        argv.extend(argv_tail);
+        let framework = classified
+            .result
+            .root
+            .as_ref()
+            .and_then(|r| r.detected_framework.clone())
+            .unwrap_or_else(|| "generic".to_string());
+
+        let mut meta = String::new();
+        meta.push_str("# Generated by `xtask audit fixtures` (corpus/README.md's workflow) —\n");
+        meta.push_str(&format!(
+            "# reviewed under seed {seed}, verdict {verdict:?}. See that file's own\n"
+        ));
+        meta.push_str(
+            "# review-any-fixture-for-machine-specific-content note before committing.\n\n",
+        );
+        // An agent generated this fixture, so `[bless] provenance` starts at
+        // the conservative default. Only a human may change this value to
+        // "human" or "agent-then-human" (corpus/README.md's `[bless]`
+        // section, the mirror of the rule that an agent must never claim
+        // `verdict_scope`) — never widen it mechanically here.
+        meta.push_str("# An agent generated this fixture; only a human may change the value\n");
+        meta.push_str("# below (corpus/README.md's `[bless]` section).\n");
+        meta.push_str("[bless]\n");
+        meta.push_str("provenance = \"agent\"\n\n");
+        meta.push_str("[tool]\n");
+        meta.push_str(&format!("name = {:?}\n", entry.tool));
+        meta.push_str(&format!("version = \"audit-seed{seed}\"\n"));
+        meta.push_str("captured_with = \"xtask audit\"\n\n");
+        meta.push_str("[[capture]]\n");
+        meta.push_str(&format!("argv = {argv:?}\n"));
+        meta.push_str("stdout = \"help.txt\"\n");
+        if !output.stderr.is_empty() {
+            meta.push_str("stderr = \"help.stderr.txt\"\n");
+        }
+        if let Some(code) = output.exit_code {
+            if code != 0 {
+                meta.push_str(&format!("exit_code = {code}\n"));
+            }
+        }
+        meta.push('\n');
+
+        match verdict {
+            "correct" => {
+                meta.push_str("[contract]\n");
+                meta.push_str(&format!("expected_framework = {framework:?}\n"));
+                if let Some(root) = classified.result.root.as_ref() {
+                    let status = status::compute(&classified.result);
+                    meta.push_str(&format!("min_status = {:?}\n", status.label));
+                    meta.push_str(&format!("min_subcommands = {}\n", root.subcommands.len()));
+                    let flags = sample_flag_specs(root);
+                    if !flags.is_empty() {
+                        meta.push_str(&format!("must_contain_flags = {flags:?}\n"));
+                    }
+                    meta.push('\n');
+                    let rendered = render_snapshot(Some(root));
+                    std::fs::write(fixture_dir.join("expected.snap"), rendered)?;
+                }
+            }
+            "incomplete" | "wrong" => {
+                meta.push_str("[contract]\n");
+                meta.push_str(&format!("expected_framework = {framework:?}\n"));
+                meta.push_str(
+                    "# TODO(human): add at least one field above (min_status/min_subcommands/\n\
+                     # must_contain_flags/must_contain_flags_by_path) that captures the specific\n\
+                     # defect the reviewer's note describes and currently FAILS against the raw\n\
+                     # capture above — xtask can't derive this without guessing at what the tool\n\
+                     # should have said, which is exactly the judgment this audit exists to add.\n\
+                     # Until then `cargo run -p xtask -- corpus` reports this fixture as \"the bug\n\
+                     # appears fixed\" (nothing here is currently falsifiable) if it's moved into\n\
+                     # a gated corpus directory — see corpus/README.md's xfail lifecycle rules.\n\n",
+                );
+                meta.push_str("[xfail]\n");
+                meta.push_str("broken = true\n");
+                let reason = if note.is_empty() {
+                    format!(
+                        "reviewer marked this {verdict} under xtask audit (seed {seed}); \
+                         no note was recorded"
+                    )
+                } else {
+                    note.to_string()
+                };
+                meta.push_str(&format!("reason = {reason:?}\n"));
+            }
+            _ => {}
+        }
+
+        std::fs::write(fixture_dir.join("meta.toml"), meta)?;
+        println!("wrote {} ({verdict})", fixture_dir.display());
+        written += 1;
+    }
+
+    println!(
+        "\n{written} fixture(s) written to {}; {skipped_verdict} skip-verdict, \
+         {skipped_exists} already existed, {skipped_no_capture} had no capture",
+        corpus_dir.display()
+    );
+    if corpus_dir != Path::new("corpus") {
+        println!(
+            "staged, not gated — review, add any needed [contract] fields to the \
+             incomplete/wrong ones, then move what's ready into corpus/ and run \
+             `cargo run -p xtask -- corpus`."
+        );
+    }
+    Ok(())
+}
+
+/// A small, generically-derived `must_contain_flags` sample for a
+/// `correct`-verdict fixture's `[contract]` — the root's first few
+/// canonical spellings, long preferred over short (matching
+/// `corpus/README.md`'s own git example). Capped rather than exhaustive:
+/// the point is a coarse regression spot-check a reviewer can extend, not a
+/// duplicate of `expected.snap`.
+pub(super) const SAMPLE_FLAG_CAP: usize = 5;
+
+pub(super) fn sample_flag_specs(root: &CommandNode) -> Vec<String> {
+    root.flags()
+        .filter_map(|f| {
+            f.long()
+                .map(|l| format!("--{l}"))
+                .or_else(|| f.short().map(|s| format!("-{s}")))
+        })
+        .take(SAMPLE_FLAG_CAP)
+        .collect()
+}
