+++ xtask/src/audit/mod.rs
@@ -35,19 +35,27 @@
 //! draw (`crate::queue::shuffle_stratify`) never consults a tool's status
 //! or name, only `(tool, stratum)` pairs and a seeded shuffle.
 
-use crate::existence::{self, FabricationKind};
+mod batch;
+mod fixtures;
+mod interactive;
+mod report;
+mod signatures;
+
+pub(crate) use batch::sanitize_filename;
+pub use batch::{cmd_amend, cmd_emit, cmd_ingest};
+pub use fixtures::cmd_fixtures;
+pub use interactive::{cmd_review, cmd_spot_audit};
+pub use report::cmd_report;
+pub(crate) use report::render_report;
+use signatures::{k1_signature, k2_signature, k3_signature};
+
+use crate::existence;
 use crate::misattribution::RecordingProbe;
 use crate::status;
-use mandible_core::audit::{
-    extract_tag_override, family_meaning, load, parse_verdict_word, save, tag_display,
-    verdict_path, AuditFile, AuditMeta, Entry,
-};
-use mandible_core::{CommandNode, Entity};
+use mandible_core::audit::Entry;
 use mandible_extract::exec::ExecOutput;
 use mandible_extract::{default_tiers_with_probe, ExtractionResult, Runner};
 use rayon::prelude::*;
-use std::collections::{BTreeMap, HashSet};
-use std::io::{BufRead, Write};
 use std::path::Path;
 use std::sync::Arc;
 
@@ -136,181 +144,6 @@ pub(crate) fn classify_all_with_recordings(
         .collect()
 }
 
-// ---------------------------------------------------------------------
-// K1/K2 pre-tagging (see [`Entry::k1`]/[`Entry::k2`]'s doc comments for the
-// full rationale). Each `_signature` function derives a suggested tag from
-// structural properties of the extracted tree/raw text alone — no tool name
-// is ever consulted, so these stay within AGENTS.md §1's "no per-tool
-// logic" invariant exactly like every other detector in this workspace.
-// ---------------------------------------------------------------------
-
-/// True for a flag matching the GCC-family single-dash-long-option defect
-/// signature: the short-flag grammar took one character as `short` and
-/// glued the rest of a multi-character single-dash spelling onto
-/// `value_name` (`-fdump-scos` -> `short=Some('f')`, `long=None`,
-/// `value_name=Some("dump-scos")`). See [`Entry::k1`].
-fn is_k1_flag(flag: &Entity) -> bool {
-    flag.short().is_some() && flag.long().is_none() && flag.value_name.is_some()
-}
-
-/// `(matching, total)` flag counts across `node` and every descendant, for
-/// the K1 pre-tag's display line (e.g. "839/1454 flags match").
-fn k1_signature_stats(node: &CommandNode) -> (usize, usize) {
-    let mut matching = node.flags().filter(|f| is_k1_flag(f)).count();
-    let mut total = node.flags().count();
-    for child in &node.subcommands {
-        let (m, t) = k1_signature_stats(child);
-        matching += m;
-        total += t;
-    }
-    (matching, total)
-}
-
-/// The K1 pre-tag suggestion: `Some(true)` when `root`'s tree contains at
-/// least one [`is_k1_flag`] match anywhere, `None` when it contains none
-/// (nothing to flag — never `Some(false)`, since there is no "confirmed not
-/// K1" state worth asserting for a tool that never exhibited the shape in
-/// the first place).
-fn k1_signature(root: &CommandNode) -> Option<bool> {
-    let (matching, _) = k1_signature_stats(root);
-    if matching > 0 {
-        Some(true)
-    } else {
-        None
-    }
-}
-
-/// True when `name` occurs as *some* whitespace-delimited token anywhere in
-/// `raw` — not restricted to a line's first token the way
-/// `existence::line_start_words` is. Punctuation immediately touching the
-/// token (as in a comma-separated list) is trimmed the same way the K2
-/// false-positive class actually presents. Used only to *explain* an
-/// existing existence-detector fabrication, never to suppress one directly
-/// — see [`k2_signature_stats`].
-fn token_occurs_anywhere(raw: &str, name: &str) -> bool {
-    raw.split_whitespace().any(|tok| {
-        tok.trim_matches(|c: char| !(c.is_alphanumeric() || c == '-' || c == '_')) == name
-    })
-}
-
-/// `(attributable, total)` counts of `report`'s subcommand-kind
-/// fabrications plausibly explained by the existence detector's own
-/// multi-column/comma-separated tokenization gap (K2) rather than genuine
-/// parser fabrication. The gap itself is closed (`existence::list_row_words`
-/// reads whole list rows), so this normally returns `(0, 0)`; kept as a
-/// regression signal.
-///
-/// Attributable: name occurs as some token anywhere in the raw text
-/// ([`token_occurs_anywhere`]), not at the line-start position the
-/// detector requires. Flag-kind fabrications are out of scope —
-/// [`existence::spelling_occurs`] already scans unconditionally.
-fn k2_signature_stats(report: &existence::ExistenceReport, raw: &str) -> (usize, usize) {
-    let subcommand_names: Vec<&str> = report
-        .fabrications
-        .iter()
-        .filter(|f| f.kind == FabricationKind::Subcommand)
-        .map(|f| f.name.as_str())
-        .collect();
-    let total = subcommand_names.len();
-    let attributable = subcommand_names
-        .iter()
-        .filter(|name| token_occurs_anywhere(raw, name))
-        .count();
-    (attributable, total)
-}
-
-/// The K2 pre-tag suggestion: `Some(true)` when *every* subcommand-kind
-/// existence fabrication for this tool is attributable to the detector's
-/// own tokenizer gap (near-certainly detector noise, not a parser defect),
-/// `Some(false)` when at least one is not (worth a real look — could be a
-/// genuine [M-10]-shaped fabrication), `None` when the tool has no
-/// subcommand-kind fabrications to judge at all.
-fn k2_signature(report: &existence::ExistenceReport, raw: &str) -> Option<bool> {
-    let (attributable, total) = k2_signature_stats(report, raw);
-    if total == 0 {
-        None
-    } else {
-        Some(attributable == total)
-    }
-}
-
-/// True for a node carrying nothing at all: no flags, no subcommands, and
-/// no summary — the same `empty` predicate `status::structure_sanity`'s own
-/// `count_suspicious` uses, reused here rather than redefined so the two
-/// "is this node genuinely empty" checks in this codebase can never drift
-/// apart.
-fn is_bare_stub(node: &CommandNode) -> bool {
-    node.flags().next().is_none() && node.subcommands.is_empty() && node.summary.is_none()
-}
-
-/// True for a bare stub ([`is_bare_stub`]) that is also not
-/// [`CommandNode::heading_attested`] — its name came from a native/cobra
-/// artifact rather than a recognized `--help` heading or headingless
-/// invocation table (spec §7 Tier B). Provable from the single extraction
-/// pass: `help_text::raw_help` refuses to probe any node whose
-/// `heading_attested` bit is false, so unlike an ordinary un-recursed
-/// subcommand, this one structurally cannot ever be probed.
-///
-/// A headingless-table node still counts here even though it's
-/// existence-attested (`invocation_attested`) — it exempts a node from
-/// being counted as *fabricated* ([`crate::status::structure_sanity`]) but
-/// does not make it any less permanently un-probed.
-///
-/// Fixture: `corpus/git-lfs/*/help.txt`.
-fn is_attestation_gated_stub(node: &CommandNode) -> bool {
-    is_bare_stub(node) && !node.heading_attested
-}
-
-/// Count of [`is_attestation_gated_stub`] matches across `node` and every
-/// descendant — called only on `root`'s subcommands, never on `root`
-/// itself, matching `status::structure_sanity`'s own root-exclusion
-/// (`root` is the literal executable name resolved from `PATH`, never
-/// something a tier guessed at, so it needs no heading to attest to).
-fn count_attestation_gated_stubs(node: &CommandNode) -> usize {
-    let this = usize::from(is_attestation_gated_stub(node));
-    this + node
-        .subcommands
-        .iter()
-        .map(count_attestation_gated_stubs)
-        .sum::<usize>()
-}
-
-/// Total flag count across `node` and every descendant.
-fn total_flags(node: &CommandNode) -> usize {
-    node.flags().count() + node.subcommands.iter().map(total_flags).sum::<usize>()
-}
-
-/// True when `root` has at least one subcommand yet the whole tree carries
-/// zero flags anywhere. `openssl`'s shape: root `--help` is a bare command
-/// grid with no options section, so the single root-only extraction pass
-/// ([`Runner::extract_full`], never recurses) surfaces no flag at all.
-/// Most tools document at least `-h`/`--version`, which is what keeps this
-/// from over-tagging every multi-level tool's ordinary lazy-fill state.
-///
-/// Fixture: `corpus/openssl/*/help.txt`.
-fn has_unfetched_subcommand_help(root: &CommandNode) -> bool {
-    !root.subcommands.is_empty() && total_flags(root) == 0
-}
-
-/// The K3 pre-tag suggestion (see [`mandible_core::audit::Entry::k3`]):
-/// `Some(true)` when `root`'s single-pass snapshot shows either known
-/// cause — an attestation-gated stub anywhere in the tree, or the
-/// whole-tree-zero-flags shape — `None` otherwise. Same "no `Some(false)`"
-/// convention as [`k1_signature`]: there is nothing to assert-not for a
-/// tool that shows neither shape.
-fn k3_signature(root: &CommandNode) -> Option<bool> {
-    let gated_stubs: usize = root
-        .subcommands
-        .iter()
-        .map(count_attestation_gated_stubs)
-        .sum();
-    if gated_stubs > 0 || has_unfetched_subcommand_help(root) {
-        Some(true)
-    } else {
-        None
-    }
-}
-
 /// Build one [`Entry`] from a classified tool, computing every pre-tag
 /// suggestion from the same single extraction pass — no second probe, same
 /// property [`Classified`]'s own doc comment describes. Shared by
@@ -380,1106 +213,15 @@ pub fn load_force_include(path: &Path) -> anyhow::Result<Vec<(String, String)>> 
     }
     Ok(out)
 }
-
-/// `xtask audit spot-audit` (spec §13.1b's sixth rule): draws `--sample`
-/// tools at random from `promoted` (the tools one mass-`ok` promotion
-/// event changed), classifies each with one fresh extraction pass
-/// ([`classify_one`]), and merges them into `<dir>/<seed>.toml` as a
-/// `spot-audit:<event>` stratum ([`effective_stratum`]).
-///
-/// The draw is reproducible: `draw_seed` mixed with `event` via
-/// [`crate::rng::stratum_seed`] seeds a Fisher-Yates shuffle
-/// ([`crate::rng::seeded_shuffle`]) over `promoted`.
-///
-/// When `promoted.len() < sample`, every promoted tool is drawn and the
-/// shortfall is stated in the summary, never silently padded.
-///
-/// A tool already present in the verdict file is tagged with
-/// [`Entry::spot_audit_event`], not duplicated or reclassified — its
-/// existing verdict is left for a human to re-review via `xtask audit
-/// amend`, never overwritten as a side effect of a draw. Re-running with
-/// the same inputs is safe: an already-tagged entry is left alone.
-pub fn cmd_spot_audit(
-    dir: &Path,
-    seed: u64,
-    event: &str,
-    promoted: &[String],
-    sample: usize,
-    draw_seed: u64,
-) -> anyhow::Result<()> {
-    if promoted.is_empty() {
-        anyhow::bail!("--promoted named no tools — nothing to spot-audit for event {event:?}");
-    }
-
-    let mut pool = promoted.to_vec();
-    crate::rng::seeded_shuffle(&mut pool, crate::rng::stratum_seed(draw_seed, event));
-    let take_n = sample.min(pool.len());
-    let drawn: Vec<String> = pool.into_iter().take(take_n).collect();
-
-    let path = verdict_path(dir, seed);
-    let mut file = if path.is_file() {
-        load(&path)?
-    } else {
-        AuditFile {
-            meta: AuditMeta {
-                seed,
-                sample_size: 0,
-            },
-            entries: Vec::new(),
-        }
-    };
-
-    let reason = if promoted.len() < sample {
-        format!(
-            "spot-audit of promotion event {event:?}: {} of {} promoted tool(s) drawn (seed \
-             {draw_seed}) — every promoted tool was audited because the promoted set was \
-             smaller than the requested sample size ({sample})",
-            drawn.len(),
-            promoted.len(),
-        )
-    } else {
-        format!(
-            "spot-audit of promotion event {event:?}: {} of {} promoted tool(s) drawn at random \
-             (seed {draw_seed})",
-            drawn.len(),
-            promoted.len(),
-        )
-    };
-
-    let existing_tools: HashSet<String> = file.entries.iter().map(|e| e.tool.clone()).collect();
-    let mut added = 0usize;
-    let mut tagged_existing = 0usize;
-    for tool in &drawn {
-        if existing_tools.contains(tool) {
-            if let Some(existing) = file.entries.iter_mut().find(|e| &e.tool == tool) {
-                if existing.spot_audit_event.is_none() {
-                    existing.spot_audit_event = Some(event.to_string());
-                    if existing.include_reason.is_none() {
-                        existing.include_reason = Some(reason.clone());
-                    }
-                    tagged_existing += 1;
-                }
-            }
-            continue;
-        }
-        let classified = classify_one(tool);
-        let mut entry = entry_from_classified(tool.clone(), &classified, Some(reason.clone()));
-        entry.spot_audit_event = Some(event.to_string());
-        file.entries.push(entry);
-        added += 1;
-    }
-    file.entries.sort_by(|a, b| a.tool.cmp(&b.tool));
-    save(&path, &file)?;
-
-    println!(
-        "spot-audit:{event}: drew {} of {} promoted tool(s)",
-        drawn.len(),
-        promoted.len(),
-    );
-    if promoted.len() < sample {
-        println!(
-            "note: the promoted set has only {} tool(s), fewer than the requested sample size \
-             {sample} — every promoted tool was audited rather than silently sampling fewer or \
-             padding the count to look like a full draw.",
-            promoted.len(),
-        );
-    }
-    println!(
-        "{added} new pending entr{s} written, {tagged_existing} already-present entr{s2} tagged \
-         into this stratum, at {} ({} tool(s) now in stratum spot-audit:{event})",
-        path.display(),
-        drawn.len(),
-        s = if added == 1 { "y" } else { "ies" },
-        s2 = if tagged_existing == 1 { "y" } else { "ies" },
-    );
-    if tagged_existing > 0 {
-        println!(
-            "note: {tagged_existing} of those were already in the file with a prior verdict — \
-             that verdict is left exactly as recorded (it may now be stale against a changed \
-             parse) for a human to re-review and correct via `xtask audit amend`, never \
-             overwritten by this draw."
-        );
-    }
-    Ok(())
-}
-
-/// Render `node` as the same YAML snapshot shown side by side with a tool's
-/// raw `--help` text in [`cmd_review`]/[`cmd_emit`] — shared so the two
-/// entry points can never render a tree differently.
-pub(crate) fn render_snapshot(node: Option<&CommandNode>) -> String {
-    match node {
-        Some(node) => {
-            let snapshot = mandible_core::to_snapshot(node);
-            serde_yaml::to_string(&snapshot)
-                .unwrap_or_else(|e| format!("(snapshot serialization failed: {e})\n"))
-        }
-        None => "(no root produced by any tier)\n".to_string(),
-    }
-}
-
-/// `xtask audit review`: the interactive loop. Presents raw `--help` text
-/// and the parsed tree for every pending entry, reads a verdict line
-/// (`<word> [note...]`) from `input`, and persists the file after every
-/// entry, so an interrupted session resumes where it stopped.
-///
-/// Line-buffered (`<word><Enter>`), not raw single-keystroke: this
-/// environment has no tty (AGENTS.md §3.2), and line buffering works
-/// identically over a real terminal or a `Cursor` in tests.
-pub fn cmd_review(
-    dir: &Path,
-    seed: u64,
-    input: &mut impl BufRead,
-    output: &mut impl Write,
-) -> anyhow::Result<()> {
-    let path = verdict_path(dir, seed);
-    let mut file = load(&path)?;
-    let pending: Vec<usize> = file.pending().collect();
-    if pending.is_empty() {
-        writeln!(output, "nothing pending in {}", path.display())?;
-        return Ok(());
-    }
-    writeln!(
-        output,
-        "{} pending of {} total. Verdict: c(orrect) / i(ncomplete) / w(rong) / s(kip), \
-         optionally followed by a space and a note. Add `k1=true`/`k1=false`/`k2=true`/\
-         `k2=false`/`k3=true`/`k3=false` anywhere in the note to override a pre-tag; omitting \
-         it confirms the suggestion shown below. Blank line or end of input stops \
-         (already-recorded verdicts are saved after every tool).",
-        pending.len(),
-        file.entries.len()
-    )?;
-
-    for idx in pending {
-        let tool = file.entries[idx].tool.clone();
-        let stratum = file.entries[idx].stratum.clone();
-        let k1 = file.entries[idx].k1;
-        let k2 = file.entries[idx].k2;
-        let k3 = file.entries[idx].k3;
-        let include_reason = file.entries[idx].include_reason.clone();
-        let classified = classify_one(&tool);
-        writeln!(output, "\n=== {tool}  (stratum: {stratum}) ===")?;
-        if let Some(reason) = &include_reason {
-            writeln!(output, "forced inclusion: {reason}")?;
-        }
-        writeln!(
-            output,
-            "{}",
-            tag_display("K1 (single-dash-long defect)", k1, "k1")
-        )?;
-        writeln!(
-            output,
-            "{}",
-            tag_display("K2 (existence-detector tokenizer gap)", k2, "k2")
-        )?;
-        writeln!(
-            output,
-            "{}",
-            tag_display("K3 (subcommand help never fetched)", k3, "k3")
-        )?;
-        writeln!(output, "--- raw --help ---")?;
-        writeln!(
-            output,
-            "{}",
-            classified
-                .raw_text
-                .as_deref()
-                .unwrap_or("(no output captured)")
-        )?;
-        writeln!(output, "--- parsed tree ---")?;
-        writeln!(
-            output,
-            "{}",
-            render_snapshot(classified.result.root.as_ref())
-        )?;
-        write!(output, "verdict> ")?;
-        output.flush()?;
-
-        let mut line = String::new();
-        let bytes_read = input.read_line(&mut line)?;
-        if bytes_read == 0 {
-            // EOF: stop here, everything already answered is already saved.
-            break;
-        }
-        let line = line.trim();
-        if line.is_empty() {
-            break;
-        }
-        let mut parts = line.splitn(2, char::is_whitespace);
-        let word = parts.next().unwrap_or("");
-        let mut note = parts.next().unwrap_or("").trim().to_string();
-        let verdict = parse_verdict_word(word)?;
-        let k1_override = extract_tag_override(&mut note, "k1");
-        let k2_override = extract_tag_override(&mut note, "k2");
-        let k3_override = extract_tag_override(&mut note, "k3");
-
-        file.entries[idx].verdict = Some(verdict.to_string());
-        file.entries[idx].note = note;
-        if let Some(v) = k1_override {
-            file.entries[idx].k1 = Some(v);
-        }
-        if let Some(v) = k2_override {
-            file.entries[idx].k2 = Some(v);
-        }
-        if let Some(v) = k3_override {
-            file.entries[idx].k3 = Some(v);
-        }
-        save(&path, &file)?;
-        writeln!(output, "recorded: {verdict}")?;
-    }
-
-    writeln!(
-        output,
-        "\n{} pending remain in {}.",
-        file.pending().count(),
-        path.display()
-    )?;
-    Ok(())
-}
-
-/// `xtask audit emit`: write every pending pair (raw text + parsed tree) to
-/// its own file under `emit_dir`, for a reviewer without a live terminal —
-/// or without this machine's tty at all — to read offline and judge on
-/// their own schedule. The counterpart, [`cmd_ingest`], reads the resulting
-/// verdicts back in.
-pub fn cmd_emit(dir: &Path, seed: u64, emit_dir: &Path) -> anyhow::Result<()> {
-    let path = verdict_path(dir, seed);
-    let file = load(&path)?;
-    std::fs::create_dir_all(emit_dir)
-        .map_err(|e| anyhow::anyhow!("creating {}: {e}", emit_dir.display()))?;
-
-    let pending: Vec<&Entry> = file
-        .entries
-        .iter()
-        .filter(|e| e.verdict.is_none())
-        .collect();
-    for entry in &pending {
-        let classified = classify_one(&entry.tool);
-        let mut buf = String::new();
-        buf.push_str(&format!(
-            "tool: {}\nstratum: {}\n",
-            entry.tool, entry.stratum
-        ));
-        if let Some(reason) = &entry.include_reason {
-            buf.push_str(&format!("forced inclusion: {reason}\n"));
-        }
-        buf.push_str(&format!(
-            "{}\n{}\n{}\n\n",
-            tag_display("K1 (single-dash-long defect)", entry.k1, "k1"),
-            tag_display("K2 (existence-detector tokenizer gap)", entry.k2, "k2"),
-            tag_display("K3 (subcommand help never fetched)", entry.k3, "k3"),
-        ));
-        buf.push_str("=== raw --help ===\n");
-        buf.push_str(
-            classified
-                .raw_text
-                .as_deref()
-                .unwrap_or("(no output captured)"),
-        );
-        buf.push_str("\n\n=== parsed tree ===\n");
-        buf.push_str(&render_snapshot(classified.result.root.as_ref()));
-        let file_path = emit_dir.join(format!("{}.txt", sanitize_filename(&entry.tool)));
-        std::fs::write(&file_path, buf)
-            .map_err(|e| anyhow::anyhow!("writing {}: {e}", file_path.display()))?;
-    }
-
-    println!(
-        "emitted {} pending pair(s) to {}",
-        pending.len(),
-        emit_dir.display()
-    );
-    println!(
-        "review offline, then write a verdicts file (one line per tool: `<tool> <verdict> \
-         [note...]`, optionally including `k1=true`/`k1=false`/`k2=true`/`k2=false`/\
-         `k3=true`/`k3=false` anywhere in the note to override a pre-tag) and run: \
-         cargo run -p xtask -- audit ingest --seed {seed} --verdicts <file>"
-    );
-    Ok(())
-}
-
-/// A tool name is never empty and, on every platform this project targets,
-/// never contains `/` (§ `resolve_tool`'s own PATH-search doesn't accept
-/// path separators in a bare tool name either), so this exists only to be
-/// defensive about the one other filesystem-hostile case worth naming.
-/// `pub(crate)` so `crate::queue`'s capture directory naming
-/// (`queue-captures/<tool>/`) can use the same rule.
-pub(crate) fn sanitize_filename(tool: &str) -> String {
-    tool.chars()
-        .map(|c| if c == '/' || c == '\\' { '_' } else { c })
-        .collect()
-}
-
-/// `xtask audit ingest`: read a plain verdicts file (`# comments` and blank
-/// lines ignored; otherwise `<tool> <verdict> [note...]` per line) and
-/// apply it to `path`'s entries. An unknown tool name is reported, not
-/// silently dropped. An entry that already carries a verdict is left alone
-/// unless `overwrite` is set — so re-running `ingest` on a file that
-/// includes already-applied lines is safe and idempotent, the same
-/// resumability property `crate::queue::cmd_sample`/[`cmd_review`] give the
-/// rest of this workflow.
-pub fn cmd_ingest(
-    dir: &Path,
-    seed: u64,
-    verdicts_path: &Path,
-    overwrite: bool,
-) -> anyhow::Result<()> {
-    let path = verdict_path(dir, seed);
-    let mut file = load(&path)?;
-    let raw = std::fs::read_to_string(verdicts_path)
-        .map_err(|e| anyhow::anyhow!("reading {}: {e}", verdicts_path.display()))?;
-
-    let mut applied = 0usize;
-    let mut already = 0usize;
-    let mut unknown: Vec<String> = Vec::new();
-
-    for (lineno, raw_line) in raw.lines().enumerate() {
-        let line = raw_line.trim();
-        if line.is_empty() || line.starts_with('#') {
-            continue;
-        }
-        let mut parts = line.splitn(3, char::is_whitespace);
-        let tool = parts.next().unwrap_or("");
-        let word = parts.next().unwrap_or("");
-        let mut note = parts.next().unwrap_or("").trim().to_string();
-        let verdict = parse_verdict_word(word)
-            .map_err(|e| anyhow::anyhow!("{}:{}: {e}", verdicts_path.display(), lineno + 1))?;
-        let k1_override = extract_tag_override(&mut note, "k1");
-        let k2_override = extract_tag_override(&mut note, "k2");
-        let k3_override = extract_tag_override(&mut note, "k3");
-
-        // The same obligation the TUI enforces, applied here so the two
-        // entry paths cannot disagree about what a complete record is. The
-        // override tokens are already stripped above, so a line whose only
-        // content was `k1=false` correctly counts as noteless.
-        if mandible_core::audit::verdict_requires_note(verdict) && note.trim().is_empty() {
-            anyhow::bail!(
-                "{}:{}: verdict {:?} for {:?} needs a note — for wrong/incomplete the note is \
-                 the finding, and an entry naming a tool with nothing about what was wrong \
-                 gives later triage nothing to act on",
-                verdicts_path.display(),
-                lineno + 1,
-                verdict,
-                tool,
-            );
-        }
-
-        let Some(entry) = file.entries.iter_mut().find(|e| e.tool == tool) else {
-            unknown.push(tool.to_string());
-            continue;
-        };
-        if entry.verdict.is_some() && !overwrite {
-            already += 1;
-            continue;
-        }
-        entry.verdict = Some(verdict.to_string());
-        entry.note = note;
-        if let Some(v) = k1_override {
-            entry.k1 = Some(v);
-        }
-        if let Some(v) = k2_override {
-            entry.k2 = Some(v);
-        }
-        if let Some(v) = k3_override {
-            entry.k3 = Some(v);
-        }
-        applied += 1;
-    }
-
-    save(&path, &file)?;
-    println!(
-        "applied {applied} verdict(s); {already} already recorded (use --overwrite to replace); \
-         {} unknown tool name(s) not in the sample{}",
-        unknown.len(),
-        if unknown.is_empty() {
-            String::new()
-        } else {
-            format!(": {}", unknown.join(", "))
-        }
-    );
-    Ok(())
-}
-
-/// `xtask audit amend`: correct one already-recorded verdict without
-/// destroying it — see `mandible_core::audit::amend`'s doc comment for the
-/// full mechanism this wraps. The only entry point touching
-/// [`mandible_core::audit::Entry::amendments`]; no TUI counterpart.
-///
-/// A subcommand, not a TUI flow: `mandible --review`'s loop
-/// (`run_review`) walks `AuditFile::needing_attention` and never revisits
-/// an already-complete `correct` verdict, so reaching an amendment needs a
-/// separate navigation mode. This command needs no tty (AGENTS.md §3.2)
-/// and is fully covered by `cargo nextest run`.
-pub fn cmd_amend(
-    dir: &Path,
-    seed: u64,
-    tool: &str,
-    new_verdict_word: &str,
-    new_note: Option<String>,
-    reason: String,
-) -> anyhow::Result<()> {
-    let path = verdict_path(dir, seed);
-    let mut file = load(&path)?;
-    let new_verdict = mandible_core::audit::parse_verdict_word(new_verdict_word)?;
-    let entry = file
-        .entries
-        .iter_mut()
-        .find(|e| e.tool == tool)
-        .ok_or_else(|| anyhow::anyhow!("{tool:?} not found in {}", path.display()))?;
-    let previous_effective = entry.effective_verdict().map(str::to_string);
-    mandible_core::audit::amend(entry, new_verdict, new_note.unwrap_or_default(), reason)?;
-    let amendment_count = entry.amendments.len();
-    save(&path, &file)?;
-    println!(
-        "amended {tool}: {} -> {new_verdict} ({amendment_count} amendment(s) now recorded for \
-         this entry)",
-        previous_effective.as_deref().unwrap_or("(none)"),
-    );
-    Ok(())
-}
-
-/// Wilson score interval for a binomial proportion at ~95% confidence
-/// (`z = 1.96`). Chosen over the naive normal approximation, which
-/// produces bounds outside `[0, 1]` at the small-`n`, near-0-or-1
-/// proportions a first audit run is likely to hit (`n=5`, `k=5`). Returns
-/// `(lower, upper)` in `[0, 1]`; `(0.0, 1.0)` for `n == 0`.
-fn wilson_interval(k: usize, n: usize) -> (f64, f64) {
-    if n == 0 {
-        return (0.0, 1.0);
-    }
-    let z = 1.96_f64;
-    let n = n as f64;
-    let p = k as f64 / n;
-    let denom = 1.0 + z * z / n;
-    let center = p + z * z / (2.0 * n);
-    let adj = z * ((p * (1.0 - p) / n) + (z * z / (4.0 * n * n))).sqrt();
-    (
-        ((center - adj) / denom).max(0.0),
-        ((center + adj) / denom).min(1.0),
-    )
-}
-
-struct StratumTally {
-    correct: usize,
-    judged: usize,
-    skipped: usize,
-    pending: usize,
-    /// Judged `wrong`/`incomplete` entries with [`Entry::is_display_only`]
-    /// true — kept out of `judged` (and therefore out of `accuracy_over`'s
-    /// denominator, spec §13.1c/task #28) but still tallied and printed,
-    /// the same "recorded, not omitted" treatment `skipped` already gets.
-    out_of_scope: usize,
-}
-
-/// The stratum label a report groups `entry` under, checked in priority
-/// order: [`Entry::spot_audit_event`] (`spot-audit:<event>`, spec §13.1b's
-/// sixth rule), else [`FORCED_INCLUSION_STRATUM`] for any other
-/// force-included entry, else the entry's own [`Entry::stratum`].
-fn effective_stratum(entry: &Entry) -> String {
-    if let Some(event) = &entry.spot_audit_event {
-        format!("spot-audit:{event}")
-    } else if entry.include_reason.is_some() {
-        FORCED_INCLUSION_STRATUM.to_string()
-    } else {
-        entry.stratum.clone()
-    }
-}
-
-/// A plain `(correct, judged)` accuracy tally over whatever subset of
-/// entries `keep` selects — shared machinery behind every accuracy number
-/// [`cmd_report`] prints.
-///
-/// Reads [`Entry::effective_verdict`], never the raw [`Entry::verdict`]:
-/// an amended entry's corrected verdict is what the project believes.
-///
-/// Also skips every [`Entry::is_display_only`] entry (task #28: a
-/// display/rendering-only finding is not an accuracy finding, though it
-/// stays a kept `wrong`/`incomplete` verdict elsewhere) — done here once
-/// rather than at each call site.
-fn accuracy_over<'a>(entries: impl Iterator<Item = &'a Entry>) -> (usize, usize) {
-    let mut correct = 0usize;
-    let mut judged = 0usize;
-    for entry in entries {
-        if entry.is_display_only() {
-            continue;
-        }
-        match entry.effective_verdict() {
-            Some("correct") => {
-                correct += 1;
-                judged += 1;
-            }
-            Some("incomplete") | Some("wrong") => judged += 1,
-            _ => {}
-        }
-    }
-    (correct, judged)
-}
-
-/// Print one `label`, count, accuracy and 95% CI line, in the shared format
-/// every accuracy line in this report uses — never a bare percentage.
-fn accuracy_line(label: &str, correct: usize, judged: usize) -> String {
-    let (lo, hi) = wilson_interval(correct, judged);
-    let acc = if judged == 0 {
-        "  n/a".to_string()
-    } else {
-        format!("{:>4.1}%", correct as f64 / judged as f64 * 100.0)
-    };
-    format!(
-        "{label:<24}  {correct:>5}/{judged:<6}  {acc}   [{:>5.1}%, {:>5.1}%]",
-        lo * 100.0,
-        hi * 100.0,
-    )
-}
-
-/// How favorable a verdict word is to the parser, for [`wilson_caveat_lines`]'s
-/// amendment-direction tally: `correct` is the best outcome, `wrong` the
-/// worst, `incomplete` between the two. `skip` has no comparable
-/// favorability (there is nothing to judge), so it is deliberately absent —
-/// an amendment into or out of `skip` is not counted as a directional move
-/// either way.
-fn verdict_favorability(verdict: &str) -> Option<i32> {
-    match verdict {
-        "correct" => Some(2),
-        "incomplete" => Some(1),
-        "wrong" => Some(0),
-        _ => None,
-    }
-}
-
-/// Print the standing caveat every accuracy figure needs: a Wilson
-/// interval bounds sampling error only, not reviewer error (every verdict
-/// is one person's unchecked read). Tallies [`Entry::amendments`] toward a
-/// more favorable outcome (`wrong`/`incomplete` -> `correct`) versus a
-/// less favorable one, via [`verdict_favorability`], and reports the
-/// actual balance rather than a hardcoded claim.
-fn wilson_caveat_lines(file: &AuditFile) -> Vec<String> {
-    let mut amended_count = 0usize;
-    let mut toward_more_favorable = 0usize;
-    let mut toward_less_favorable = 0usize;
-    for entry in &file.entries {
-        if !entry.amendments.is_empty() {
-            amended_count += 1;
-        }
-        for amendment in &entry.amendments {
-            if let (Some(before), Some(after)) = (
-                verdict_favorability(&amendment.previous_verdict),
-                verdict_favorability(&amendment.new_verdict),
-            ) {
-                match after.cmp(&before) {
-                    std::cmp::Ordering::Greater => toward_more_favorable += 1,
-                    std::cmp::Ordering::Less => toward_less_favorable += 1,
-                    std::cmp::Ordering::Equal => {}
-                }
-            }
-        }
-    }
-    let mut lines = vec![
-        String::new(),
-        "note: the 95% CI above bounds sampling error only — how much this sample's accuracy \
-         could plausibly vary on a fresh draw of the same size — never reviewer error. Read the \
-         accuracy figure as \"accuracy of the parser as judged by this reviewer,\" not an \
-         absolute truth."
-            .to_string(),
-    ];
-    if amended_count == 0 {
-        lines.push(
-            "note: no verdict in this file has been amended yet (`xtask audit amend`) — this \
-             says nothing about whether the recorded verdicts are all correct, only that none \
-             has been corrected so far."
-                .to_string(),
-        );
-    } else {
-        lines.push(format!(
-            "note: {amended_count} verdict(s) carry a recorded amendment; of the corrections \
-             with a comparable direction, {toward_less_favorable} made the verdict less \
-             favorable to the parser (an originally too-generous read) and \
-             {toward_more_favorable} made it more favorable (an originally too-harsh read).{}",
-            if toward_less_favorable > toward_more_favorable {
-                " More corrections have gone the generous-to-harsh direction than the reverse \
-                 so far, so this accuracy figure likely still reads a little high."
-            } else if toward_more_favorable > toward_less_favorable {
-                " More corrections have gone the harsh-to-generous direction than the reverse \
-                 so far, so this accuracy figure likely still reads a little low."
-            } else {
-                " The corrections so far do not lean toward either direction."
-            }
-        ));
-    }
-    lines
-}
-
-/// The `skip` verdicts, named — the stratum table prints only a `skipped`
-/// count per stratum, which makes the accuracy denominator's exclusion
-/// unauditable on its own. `skip` is recorded, not omitted (spec §16):
-/// every skipped tool by name, with the reviewer's reason or an explicit
-/// `(no reason recorded)`.
-///
-/// Returns whole lines (header included) rather than printing, so content
-/// is testable without capturing stdout.
-fn skipped_lines(file: &AuditFile) -> Vec<String> {
-    let mut skipped: Vec<&Entry> = file
-        .entries
-        .iter()
-        .filter(|e| e.effective_verdict() == Some("skip"))
-        .collect();
-    if skipped.is_empty() {
-        return Vec::new();
-    }
-    skipped.sort_by(|a, b| a.tool.cmp(&b.tool));
-    let mut lines = vec![
-        String::new(),
-        format!(
-            "tools skipped ({} — recorded, never omitted; excluded from every accuracy figure \
-         above, so this is the list that makes that exclusion checkable):",
-            skipped.len()
-        ),
-    ];
-    for entry in skipped {
-        let reason = if entry.effective_note().trim().is_empty() {
-            "(no reason recorded)"
-        } else {
-            entry.effective_note()
-        };
-        lines.push(format!("  {:<24} {}", entry.tool, reason));
-    }
-    lines
-}
-
-/// Build `xtask audit report`'s full text without printing it, so a caller
-/// (`cmd_report` itself, and `xtask audit contribute`'s `<seed>-report.txt`)
-/// can both use exactly the same rendering rather than one re-deriving it or
-/// scraping the other's stdout (AGENTS.md §3.3: never parse human-format
-/// output, including your own).
-// Ratchet: one report template rendered in a single pass. Listed in scripts/ratchet.txt.
-#[allow(clippy::too_many_lines)]
-pub(crate) fn render_report(dir: &Path, seed: u64) -> anyhow::Result<String> {
-    let path = verdict_path(dir, seed);
-    let file = load(&path)?;
-
-    let mut by_stratum: BTreeMap<String, StratumTally> = BTreeMap::new();
-    for entry in &file.entries {
-        let tally = by_stratum
-            .entry(effective_stratum(entry))
-            .or_insert(StratumTally {
-                correct: 0,
-                judged: 0,
-                skipped: 0,
-                pending: 0,
-                out_of_scope: 0,
-            });
-        match entry.effective_verdict() {
-            None => tally.pending += 1,
-            Some("skip") => tally.skipped += 1,
-            Some("correct") => {
-                tally.correct += 1;
-                tally.judged += 1;
-            }
-            // A display-only finding is judged (`wrong`/`incomplete`, never
-            // `skip` — see `Entry::is_display_only`'s doc comment on why
-            // `skip` is the wrong tool for this), so it must not fall into
-            // the catch-all `judged` arm below: that is precisely the
-            // count `accuracy_over` also excludes it from. Checked before
-            // the catch-all, not after, so it can never double-count.
-            Some(_) if entry.is_display_only() => tally.out_of_scope += 1,
-            Some(_) => tally.judged += 1,
-        }
-    }
-
-    let mut lines: Vec<String> = Vec::new();
-    lines.push(format!(
-        "audit seed={seed} sample_size={} ({} entries total)",
-        file.meta.sample_size,
-        file.entries.len()
-    ));
-    lines.push(String::new());
-    lines.push(
-        "stratum             correct/judged   accuracy   95% CI            skipped   pending   \
-         out-of-scope"
-            .to_string(),
-    );
-    let mut overall_correct = 0usize;
-    let mut overall_judged = 0usize;
-    let mut overall_skipped = 0usize;
-    let mut overall_pending = 0usize;
-    let mut overall_out_of_scope = 0usize;
-    for (stratum, t) in &by_stratum {
-        let (lo, hi) = wilson_interval(t.correct, t.judged);
-        let acc = if t.judged == 0 {
-            "  n/a".to_string()
-        } else {
-            format!("{:>4.1}%", t.correct as f64 / t.judged as f64 * 100.0)
-        };
-        lines.push(format!(
-            "{stratum:<18}  {:>5}/{:<6}  {acc}   [{:>5.1}%, {:>5.1}%]   {:>7}   {:>7}   {:>12}",
-            t.correct,
-            t.judged,
-            lo * 100.0,
-            hi * 100.0,
-            t.skipped,
-            t.pending,
-            t.out_of_scope,
-        ));
-        overall_correct += t.correct;
-        overall_judged += t.judged;
-        overall_skipped += t.skipped;
-        overall_pending += t.pending;
-        overall_out_of_scope += t.out_of_scope;
-    }
-    let (lo, hi) = wilson_interval(overall_correct, overall_judged);
-    let overall_acc = if overall_judged == 0 {
-        "  n/a".to_string()
-    } else {
-        format!(
-            "{:>4.1}%",
-            overall_correct as f64 / overall_judged as f64 * 100.0
-        )
-    };
-    lines.push(format!(
-        "{:<18}  {:>5}/{:<6}  {overall_acc}   [{:>5.1}%, {:>5.1}%]   {:>7}   {:>7}   {:>12}",
-        "OVERALL",
-        overall_correct,
-        overall_judged,
-        lo * 100.0,
-        hi * 100.0,
-        overall_skipped,
-        overall_pending,
-        overall_out_of_scope,
-    ));
-    if overall_judged > 0 && overall_judged < 30 {
-        lines.push(format!(
-            "\nnote: n={overall_judged} judged so far — the interval above is wide at this size; \
-             keep reviewing for a number worth acting on (spec's own target is ~60-100)."
-        ));
-    }
-    if overall_out_of_scope > 0 {
-        let mut names: Vec<&str> = file
-            .entries
-            .iter()
-            .filter(|e| e.is_display_only())
-            .map(|e| e.tool.as_str())
-            .collect();
-        names.sort_unstable();
-        lines.push(format!(
-            "\nnote: {overall_out_of_scope} finding(s) are display-only and are excluded from \
-             every accuracy figure above, not dropped — the maintainer's ruling (task #28) is \
-             that a display/rendering defect is a real finding but not an accuracy one: {}. See \
-             the 'display-only findings (kept, out of scope)' section below for each one's note \
-             in full.",
-            names.join(", "),
-        ));
-    }
-    lines.extend(wilson_caveat_lines(&file));
-
-    let k1_tagged = file.entries.iter().filter(|e| e.k1 == Some(true)).count();
-    let k2_tagged = file.entries.iter().filter(|e| e.k2 == Some(true)).count();
-    let k3_tagged = file.entries.iter().filter(|e| e.k3 == Some(true)).count();
-    lines.push(format!(
-        "\nK1/K2/K3 sensitivity ({k1_tagged} entr{k1_s} tagged K1, {k2_tagged} entr{k2_s} \
-         tagged K2, {k3_tagged} entr{k3_s} tagged K3 — see mandible_core::audit's \
-         Entry::k1/k2/k3 doc comments and this module's *_signature functions):",
-        k1_s = if k1_tagged == 1 { "y" } else { "ies" },
-        k2_s = if k2_tagged == 1 { "y" } else { "ies" },
-        k3_s = if k3_tagged == 1 { "y" } else { "ies" },
-    ));
-    lines.push("view                      correct/judged   accuracy   95% CI".to_string());
-    let (c, j) = accuracy_over(file.entries.iter());
-    lines.push(accuracy_line("all-inclusive", c, j));
-    let (c, j) = accuracy_over(file.entries.iter().filter(|e| e.k1 != Some(true)));
-    lines.push(accuracy_line("K1-excluded", c, j));
-    let (c, j) = accuracy_over(file.entries.iter().filter(|e| e.k2 != Some(true)));
-    lines.push(accuracy_line("K2-excluded", c, j));
-    let (c, j) = accuracy_over(file.entries.iter().filter(|e| e.k3 != Some(true)));
-    lines.push(accuracy_line("K3-excluded", c, j));
-    let (c, j) = accuracy_over(
-        file.entries
-            .iter()
-            .filter(|e| e.k1 != Some(true) && e.k2 != Some(true) && e.k3 != Some(true)),
-    );
-    lines.push(accuracy_line("K1+K2+K3-excluded", c, j));
-
-    let mut flagged: Vec<&Entry> = file
-        .entries
-        .iter()
-        .filter(|e| matches!(e.effective_verdict(), Some("wrong") | Some("incomplete")))
-        .collect();
-    flagged.sort_by(|a, b| a.tool.cmp(&b.tool));
-    if !flagged.is_empty() {
-        lines.push("\ntools judged wrong or incomplete (the next bugs):".to_string());
-        for entry in flagged {
-            let amended_tag = if entry.amendments.is_empty() {
-                ""
-            } else {
-                " [amended]"
-            };
-            // Stays in this list — it is still a `wrong`/`incomplete`
-            // verdict on disk and a real finding — but tagged so a reader
-            // scanning "the next bugs" does not mistake a rendering fix
-            // for a parser fix. `accuracy_over` has already excluded it
-            // from every count printed above; this tag is why the two
-            // views (this list and the headline) don't silently disagree
-            // about which tools are counted where.
-            let scope_tag = if entry.is_display_only() {
-                " [display-only, excluded from accuracy — see below]"
-            } else {
-                ""
-            };
-            lines.push(format!(
-                "  {:<24} {:<11} {}{amended_tag}{scope_tag}",
-                entry.tool,
-                entry.effective_verdict().unwrap_or(""),
-                entry.effective_note(),
-            ));
-        }
-    }
-
-    lines.extend(skipped_lines(&file));
-
-    let mut out_of_scope: Vec<&Entry> = file
-        .entries
-        .iter()
-        .filter(|e| e.is_display_only())
-        .collect();
-    out_of_scope.sort_by(|a, b| a.tool.cmp(&b.tool));
-    if !out_of_scope.is_empty() {
-        lines.push(format!(
-            "\ndisplay-only findings (kept, out of scope — real UI bugs, excluded from accuracy \
-             per the maintainer's task #28 ruling; family meaning: {}):",
-            family_meaning("display-only").unwrap_or("?"),
-        ));
-        for entry in out_of_scope {
-            lines.push(format!(
-                "  {:<24} {:<11} {}",
-                entry.tool,
-                entry.effective_verdict().unwrap_or(""),
-                entry.effective_note(),
-            ));
-        }
-    }
-    lines.push(String::new());
-    Ok(lines.join("\n"))
-}
-
-/// `xtask audit report`: per-stratum and overall accuracy as a count and
-/// confidence interval, never a bare percentage (spec §13.1b). Also
-/// reports four K1/K2 views (all-inclusive, K1-excluded, K2-excluded,
-/// both-excluded — see [`Entry::k1`]/[`Entry::k2`]) and lists every tool
-/// judged `wrong`/`incomplete`.
-pub fn cmd_report(dir: &Path, seed: u64) -> anyhow::Result<()> {
-    print!("{}", render_report(dir, seed)?);
-    Ok(())
-}
-
-/// `xtask audit fixtures`: turn every reviewed (non-`skip`) entry into a
-/// `corpus/README.md`-shaped fixture directory under `corpus_dir` — capture
-/// files, a pre-filled `meta.toml`, and (for `correct`) an `expected.snap`.
-///
-/// Stages by default (`corpus_dir` defaults to `<dir>/<seed>/fixtures`,
-/// not `corpus/`): a `wrong`/`incomplete` verdict becomes an `[xfail]`
-/// with no currently-failing `[contract]` field, which `xtask corpus`
-/// otherwise reads as "the bug appears fixed" and vacuously passes.
-/// Writes the reviewer's note as `reason` plus `expected_framework`
-/// (Tier A′'s detection, not a correctness claim). Promoting a staged
-/// fixture into `corpus/` is a deliberate act, same spirit as `--bless`.
-///
-/// A `correct` verdict gets a real `expected.snap` and may ship green
-/// immediately, wherever `corpus_dir` points.
-pub fn cmd_fixtures(
-    dir: &Path,
-    seed: u64,
-    corpus_dir: &Path,
-    only: Option<Vec<String>>,
-    force: bool,
-) -> anyhow::Result<()> {
-    let path = verdict_path(dir, seed);
-    let file = load(&path)?;
-    let mut written = 0usize;
-    let mut skipped_no_capture = 0usize;
-    let mut skipped_verdict = 0usize;
-    let mut skipped_exists = 0usize;
-
-    for entry in &file.entries {
-        // Reads the effective (post-amendment) verdict and note: a fixture
-        // generated from an entry that was amended after review must
-        // reflect the corrected truth, not the superseded original — the
-        // whole point of an amendment is that the project's belief about
-        // the tool changed, and a fixture is exactly the kind of durable
-        // artifact that would otherwise silently encode the old, wrong
-        // belief forever.
-        let Some(verdict) = entry.effective_verdict() else {
-            continue;
-        };
-        let note = entry.effective_note();
-        if verdict == "skip" {
-            skipped_verdict += 1;
-            continue;
-        }
-        if let Some(only) = &only {
-            if !only.iter().any(|t| t == &entry.tool) {
-                continue;
-            }
-        }
-
-        let fixture_dir = corpus_dir
-            .join(&entry.tool)
-            .join(format!("audit-seed{seed}"));
-        if fixture_dir.exists() && !force {
-            println!(
-                "{}: {} already exists — pass --force to overwrite (review the existing \
-                 fixture for machine-specific content first, corpus/README.md step 3)",
-                entry.tool,
-                fixture_dir.display()
-            );
-            skipped_exists += 1;
-            continue;
-        }
-
-        let classified = classify_one(&entry.tool);
-        let Some((argv_tail, output)) = classified.raw_capture else {
-            println!(
-                "{}: no raw capture available, skipping fixture emission",
-                entry.tool
-            );
-            skipped_no_capture += 1;
-            continue;
-        };
-
-        std::fs::create_dir_all(&fixture_dir)
-            .map_err(|e| anyhow::anyhow!("creating {}: {e}", fixture_dir.display()))?;
-        std::fs::write(fixture_dir.join("help.txt"), &output.stdout)?;
-        if !output.stderr.is_empty() {
-            std::fs::write(fixture_dir.join("help.stderr.txt"), &output.stderr)?;
-        }
-
-        let mut argv = vec![entry.tool.clone()];
-        argv.extend(argv_tail);
-        let framework = classified
-            .result
-            .root
-            .as_ref()
-            .and_then(|r| r.detected_framework.clone())
-            .unwrap_or_else(|| "generic".to_string());
-
-        let mut meta = String::new();
-        meta.push_str("# Generated by `xtask audit fixtures` (corpus/README.md's workflow) —\n");
-        meta.push_str(&format!(
-            "# reviewed under seed {seed}, verdict {verdict:?}. See that file's own\n"
-        ));
-        meta.push_str(
-            "# review-any-fixture-for-machine-specific-content note before committing.\n\n",
-        );
-        // An agent generated this fixture, so `[bless] provenance` starts at
-        // the conservative default. Only a human may change this value to
-        // "human" or "agent-then-human" (corpus/README.md's `[bless]`
-        // section, the mirror of the rule that an agent must never claim
-        // `verdict_scope`) — never widen it mechanically here.
-        meta.push_str("# An agent generated this fixture; only a human may change the value\n");
-        meta.push_str("# below (corpus/README.md's `[bless]` section).\n");
-        meta.push_str("[bless]\n");
-        meta.push_str("provenance = \"agent\"\n\n");
-        meta.push_str("[tool]\n");
-        meta.push_str(&format!("name = {:?}\n", entry.tool));
-        meta.push_str(&format!("version = \"audit-seed{seed}\"\n"));
-        meta.push_str("captured_with = \"xtask audit\"\n\n");
-        meta.push_str("[[capture]]\n");
-        meta.push_str(&format!("argv = {argv:?}\n"));
-        meta.push_str("stdout = \"help.txt\"\n");
-        if !output.stderr.is_empty() {
-            meta.push_str("stderr = \"help.stderr.txt\"\n");
-        }
-        if let Some(code) = output.exit_code {
-            if code != 0 {
-                meta.push_str(&format!("exit_code = {code}\n"));
-            }
-        }
-        meta.push('\n');
-
-        match verdict {
-            "correct" => {
-                meta.push_str("[contract]\n");
-                meta.push_str(&format!("expected_framework = {framework:?}\n"));
-                if let Some(root) = classified.result.root.as_ref() {
-                    let status = status::compute(&classified.result);
-                    meta.push_str(&format!("min_status = {:?}\n", status.label));
-                    meta.push_str(&format!("min_subcommands = {}\n", root.subcommands.len()));
-                    let flags = sample_flag_specs(root);
-                    if !flags.is_empty() {
-                        meta.push_str(&format!("must_contain_flags = {flags:?}\n"));
-                    }
-                    meta.push('\n');
-                    let rendered = render_snapshot(Some(root));
-                    std::fs::write(fixture_dir.join("expected.snap"), rendered)?;
-                }
-            }
-            "incomplete" | "wrong" => {
-                meta.push_str("[contract]\n");
-                meta.push_str(&format!("expected_framework = {framework:?}\n"));
-                meta.push_str(
-                    "# TODO(human): add at least one field above (min_status/min_subcommands/\n\
-                     # must_contain_flags/must_contain_flags_by_path) that captures the specific\n\
-                     # defect the reviewer's note describes and currently FAILS against the raw\n\
-                     # capture above — xtask can't derive this without guessing at what the tool\n\
-                     # should have said, which is exactly the judgment this audit exists to add.\n\
-                     # Until then `cargo run -p xtask -- corpus` reports this fixture as \"the bug\n\
-                     # appears fixed\" (nothing here is currently falsifiable) if it's moved into\n\
-                     # a gated corpus directory — see corpus/README.md's xfail lifecycle rules.\n\n",
-                );
-                meta.push_str("[xfail]\n");
-                meta.push_str("broken = true\n");
-                let reason = if note.is_empty() {
-                    format!(
-                        "reviewer marked this {verdict} under xtask audit (seed {seed}); \
-                         no note was recorded"
-                    )
-                } else {
-                    note.to_string()
-                };
-                meta.push_str(&format!("reason = {reason:?}\n"));
-            }
-            _ => {}
-        }
-
-        std::fs::write(fixture_dir.join("meta.toml"), meta)?;
-        println!("wrote {} ({verdict})", fixture_dir.display());
-        written += 1;
-    }
-
-    println!(
-        "\n{written} fixture(s) written to {}; {skipped_verdict} skip-verdict, \
-         {skipped_exists} already existed, {skipped_no_capture} had no capture",
-        corpus_dir.display()
-    );
-    if corpus_dir != Path::new("corpus") {
-        println!(
-            "staged, not gated — review, add any needed [contract] fields to the \
-             incomplete/wrong ones, then move what's ready into corpus/ and run \
-             `cargo run -p xtask -- corpus`."
-        );
-    }
-    Ok(())
-}
-
-/// A small, generically-derived `must_contain_flags` sample for a
-/// `correct`-verdict fixture's `[contract]` — the root's first few
-/// canonical spellings, long preferred over short (matching
-/// `corpus/README.md`'s own git example). Capped rather than exhaustive:
-/// the point is a coarse regression spot-check a reviewer can extend, not a
-/// duplicate of `expected.snap`.
-const SAMPLE_FLAG_CAP: usize = 5;
-
-fn sample_flag_specs(root: &CommandNode) -> Vec<String> {
-    root.flags()
-        .filter_map(|f| {
-            f.long()
-                .map(|l| format!("--{l}"))
-                .or_else(|| f.short().map(|s| format!("-{s}")))
-        })
-        .take(SAMPLE_FLAG_CAP)
-        .collect()
-}
-
 #[cfg(test)]
 mod tests {
+    use super::report::*;
+    use super::signatures::*;
     use super::*;
-    use mandible_core::audit::AuditMeta;
-    use mandible_core::{Provenance, Source};
+    use mandible_core::audit::{
+        extract_tag_override, load, parse_verdict_word, save, verdict_path, AuditFile, AuditMeta,
+    };
+    use mandible_core::{CommandNode, Entity, Provenance, Source};
     use std::io::Cursor;
     use std::path::PathBuf;
